Serialization dependency removed
