--- test/display_serialization_test.xml
@@ -1,34 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="yes" ?>
-<!DOCTYPE boost_serialization>
-<boost_serialization signature="serialization::archive" version="17">
-<item class_id="0" tracking_level="0" version="0">
-	<axes class_id="1" tracking_level="0" version="0">
-		<item class_id="2" tracking_level="0" version="0">
-			<transform class_id="3" tracking_level="0" version="0"></transform>
-			<size>10</size>
-			<meta></meta>
-			<min>-3.00000000000000000e+00</min>
-			<delta>6.00000000000000000e+00</delta>
-		</item>
-	</axes>
-	<storage class_id="4" tracking_level="0" version="0">
-		<type>1</type>
-		<size>12</size>
-		<buffer>
-			<item>0</item>
-			<item>819</item>
-			<item>1181</item>
-			<item>854</item>
-			<item>1162</item>
-			<item>884</item>
-			<item>1037</item>
-			<item>1020</item>
-			<item>1203</item>
-			<item>844</item>
-			<item>996</item>
-			<item>0</item>
-		</buffer>
-	</storage>
-</item>
-</boost_serialization>
-
